requirements.txt
amqp==2.6.1
anyjson==0.3.3
appnope==0.1.0
asgiref==3.2.10
Babel==2.8.0
backcall==0.2.0
billiard==3.6.3.0
celery==4.3.0
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
decorator==4.4.2
Django==3.1.2
django-mysql==3.9.0
djangorestframework==3.12.1
flower==0.9.3
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
kombu==4.6.11
mysqlclient==2.0.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.2
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
tornado==5.1.1
traitlets==5.0.5
vine==1.3.0
wcwidth==0.2.5
